Rename project to cpex (CSP Process Explorer).
System for exporting to C now working, and should be system-independent,
to a reasonable extent. That is, assuming that the person compiling the
software is using cabal to build haskell packages and has built the
shared libraries.

A script (which-hs) is needed to assist the Makefile during compilation,
since Haskell is a bit awkward about where it wants the .so files.