Updated build system to make use of the gnu build tools.
Cross platform ahoy! To build:

$ autoreconf -i
$ ./configure
$ make

Also removed which-hs. That went obsolete quickly! It now exists in the
configure script.

TODO: "cabal build" for the haskell component currently happens in the
configure step. This is bad! Things are mostly complicated by the fact
that the .so library could be called one of many things.

TODO: Get make install to put libHScpex somewhere, or rely on cabal.
Decide what to do here.