Changes to model structure.
Factors out the export functions in haskell to their own file. Processes
are now separated from the item they represent in the model to allow for
other representations (currently a graph, with the model using a tree to
allow for traces). ProcessItem and SessionItem can be found as private
classes in the respective Model cpp files.