Improvements to HTML rendering
String replacement changed to be handled by a DisplayString class, which
avoids having to run the same code on the same input multiple times.
Also added a sizeHint to correctly determine the desired column width,
instead of using the text width of the raw HTML (which is quite a bit
wider).