Stops prettyPrint output from being used as input
Now the actual underlying process is used when probe/inspect is chosen
from the context menu, which is a more robust method and doesn't rely on
the prettyPrint output being unambiguous. This involves the inclusion of
a new base class ProcessItem from which TransitionItem and InspectItem
derive; a good thing to have anyway to avoid code duplication.