Implements omega semantics
The transitions function now accepts an extra argument: the process that
should be used as Omega (defined in the session), the process that has
already terminated (distinct from but semantically equivalent to STOP).
If not given, synchronous termination semantics are used instead, and
STOP is the result of performing a Tick.