Finishes implementation of error handling
Errors now show up in the transitions tree when an erroneous process
appears, if the root process is lazily deemed to be valid. If not, then
the transitions tree does not show up at all. To achieve this, an "error
process" type is added, which stores error messages associated with that
process.