Adds a context menu to session listing.
It is now possible to launch the inspector directly from the session
listing without having to type directly into the box, using the context
menu.

A hidden bug in the tab code, where updateExprBox would fail when the
current tab is blank, has been revealed by this change and fixed.