Complete separation of model from view
ProgramState is now a QObject rather than a series of static methods,
removing its reliance on MainWindow in favour of signals and slots. Can
be accessed by ProgramState::get(), which returns a reference to the
singleton ProgramState object.

Everywhere else in the model, if there is a reliance on some widget
class or on MainWindow, the code is reworked to use signals and slots
instead. In most instances this was already the case.